Refactor Apps components
